+++ src/main/java/com/github/money/keeper/storage/AccountRepo.java
@@ -2,5 +2,5 @@ package com.github.money.keeper.storage;
 
 import com.github.money.keeper.model.Account;
 
-public interface AccountRepo extends BaseRepo<Long, Account> {
+public interface AccountRepo extends BaseRepo<Integer, Account> {
 }
